move inputs into more modules
