+++ convex/game/conversationMembers.ts
@@ -5,6 +5,7 @@ import { DatabaseReader, DatabaseWriter } from '../_generated/server';
 import { Doc, Id } from '../_generated/dataModel';
 import { Conversations, stopConversation } from './conversations';
 import { AiTown } from './aiTown';
+import { inputHandler } from './inputs';
 
 export const conversationMembers = defineTable({
   conversationId: v.id('conversations'),
@@ -143,3 +144,42 @@ export function leaveConversation(
   }
   stopConversation(game, now, conversation);
 }
+
+export const conversationMembersInputs = {
+  // Accept an invite to a conversation, which puts the
+  // player in the "walkingOver" state until they're close
+  // enough to the other participant.
+  acceptInvite: inputHandler({
+    args: {
+      playerId: v.id('players'),
+      conversationId: v.id('conversations'),
+    },
+    handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
+      acceptInvite(game, playerId, conversationId);
+      return null;
+    },
+  }),
+  // Reject the invite. Eventually we might add a message
+  // that explains why!
+  rejectInvite: inputHandler({
+    args: {
+      playerId: v.id('players'),
+      conversationId: v.id('conversations'),
+    },
+    handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
+      rejectInvite(game, now, playerId, conversationId);
+      return null;
+    },
+  }),
+  // Leave a conversation.
+  leaveConversation: inputHandler({
+    args: {
+      playerId: v.id('players'),
+      conversationId: v.id('conversations'),
+    },
+    handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
+      leaveConversation(game, now, playerId, conversationId);
+      return null;
+    },
+  }),
+};
