+++ convex/game/conversations.ts
@@ -4,6 +4,7 @@ import { GameTable } from '../engine/gameTable';
 import { DatabaseWriter } from '../_generated/server';
 import { Doc, Id } from '../_generated/dataModel';
 import { AiTown } from './aiTown';
+import { inputHandler } from './inputs';
 
 export const conversations = defineTable({
   worldId: v.id('worlds'),
@@ -133,3 +134,84 @@ export function stopConversation(game: AiTown, now: number, conversation: Doc<'c
     }
   }
 }
+
+// Start a conversation, inviting the specified player.
+// Conversations can only have two participants for now,
+// so we don't have a separate "invite" input.
+const startConversationInput = inputHandler({
+  args: {
+    playerId: v.id('players'),
+    invitee: v.id('players'),
+  },
+  handler: async (
+    game: AiTown,
+    now: number,
+    { playerId, invitee },
+  ): Promise<Id<'conversations'>> => {
+    console.log(`Starting ${playerId} ${invitee}...`);
+    const { conversationId, error } = await startConversation(game, playerId, invitee);
+    if (!conversationId) {
+      // TODO: pass it back to the client for them to show an error.
+      throw new Error(error);
+    }
+    return conversationId;
+  },
+});
+
+const startTyping = inputHandler({
+  args: {
+    playerId: v.id('players'),
+    conversationId: v.id('conversations'),
+    messageUuid: v.string(),
+  },
+  handler: async (
+    game: AiTown,
+    now: number,
+    { playerId, conversationId, messageUuid },
+  ): Promise<null> => {
+    const conversation = game.conversations.lookup(conversationId);
+    if (!conversation) {
+      throw new Error(`Invalid conversation ID: ${conversationId}`);
+    }
+    if (conversation.isTyping && conversation.isTyping.playerId !== playerId) {
+      throw new Error(
+        `Player ${conversation.isTyping.playerId} is already typing in ${conversationId}`,
+      );
+    }
+    conversation.isTyping = { playerId, messageUuid, since: now };
+    return null;
+  },
+});
+
+export const finishSendingMessage = inputHandler({
+  args: {
+    playerId: v.id('players'),
+    conversationId: v.id('conversations'),
+    timestamp: v.number(),
+  },
+  handler: async (
+    game: AiTown,
+    now: number,
+    { playerId, conversationId, timestamp },
+  ): Promise<null> => {
+    const conversation = game.conversations.lookup(conversationId);
+    if (!conversation) {
+      throw new Error(`Invalid conversation ID: ${conversationId}`);
+    }
+    if (conversation.finished) {
+      throw new Error(`Conversation is finished: ${conversationId}`);
+    }
+    if (conversation.isTyping && conversation.isTyping.playerId === playerId) {
+      delete conversation.isTyping;
+    }
+    conversation.lastMessage = { author: playerId, timestamp };
+    conversation.numMessages++;
+    return null;
+  },
+});
+
+export const conversationInputs = {
+  startConversation: startConversationInput,
+  startTyping,
+  finishSendingMessage,
+};
