+++ convex/game/inputs.ts
@@ -1,11 +1,10 @@
-import { Id } from '../_generated/dataModel';
 import { blocked } from './movement';
 import { characters } from '../../data/characters';
 import { agentInputs } from './agents';
 import { ObjectType, PropertyValidators, v } from 'convex/values';
 import { movePlayer } from './movement';
-import * as conversationMembers from './conversationMembers';
-import * as conversations from './conversations';
+import { conversationMembersInputs } from './conversationMembers';
+import { conversationInputs, stopConversation } from './conversations';
 import { point } from '../util/types';
 import { AiTown } from './aiTown';
 
@@ -70,7 +69,7 @@ export const join = inputHandler({
     });
     return playerId;
   },
-} as const);
+});
 // ...or leave, disabling the specified player.
 const leave = inputHandler({
   args: {
@@ -85,12 +84,12 @@ const leave = inputHandler({
       if (conversation === null) {
         throw new Error(`Couldn't find conversation: ${membership.conversationId}`);
       }
-      conversations.stopConversation(game, now, conversation);
+      stopConversation(game, now, conversation);
     }
     player.active = false;
     return null;
   },
-} as const);
+});
 // Move the player to a specified location.
 const moveTo = inputHandler({
   args: {
@@ -102,131 +101,13 @@ const moveTo = inputHandler({
     return null;
   },
 });
-// Start a conversation, inviting the specified player.
-// Conversations can only have two participants for now,
-// so we don't have a separate "invite" input.
-const startConversation = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    invitee: v.id('players'),
-  },
-  handler: async (
-    game: AiTown,
-    now: number,
-    { playerId, invitee },
-  ): Promise<Id<'conversations'>> => {
-    console.log(`Starting ${playerId} ${invitee}...`);
-    const { conversationId, error } = await conversations.startConversation(
-      game,
-      playerId,
-      invitee,
-    );
-    if (!conversationId) {
-      // TODO: pass it back to the client for them to show an error.
-      throw new Error(error);
-    }
-    return conversationId;
-  },
-});
-// Accept an invite to a conversation, which puts the
-// player in the "walkingOver" state until they're close
-// enough to the other participant.
-const acceptInvite = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    conversationId: v.id('conversations'),
-  },
-  handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
-    conversationMembers.acceptInvite(game, playerId, conversationId);
-    return null;
-  },
-});
-// Reject the invite. Eventually we might add a message
-// that explains why!
-const rejectInvite = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    conversationId: v.id('conversations'),
-  },
-  handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
-    conversationMembers.rejectInvite(game, now, playerId, conversationId);
-    return null;
-  },
-});
-// Leave a conversation.
-const leaveConversation = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    conversationId: v.id('conversations'),
-  },
-  handler: async (game: AiTown, now: number, { playerId, conversationId }): Promise<null> => {
-    conversationMembers.leaveConversation(game, now, playerId, conversationId);
-    return null;
-  },
-});
-// Inputs for the messaging layer.
-const startTyping = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    conversationId: v.id('conversations'),
-    messageUuid: v.string(),
-  },
-  handler: async (
-    game: AiTown,
-    now: number,
-    { playerId, conversationId, messageUuid },
-  ): Promise<null> => {
-    const conversation = game.conversations.lookup(conversationId);
-    if (!conversation) {
-      throw new Error(`Invalid conversation ID: ${conversationId}`);
-    }
-    if (conversation.isTyping && conversation.isTyping.playerId !== playerId) {
-      throw new Error(
-        `Player ${conversation.isTyping.playerId} is already typing in ${conversationId}`,
-      );
-    }
-    conversation.isTyping = { playerId, messageUuid, since: now };
-    return null;
-  },
-});
-
-export const finishSendingMessage = inputHandler({
-  args: {
-    playerId: v.id('players'),
-    conversationId: v.id('conversations'),
-    timestamp: v.number(),
-  },
-  handler: async (
-    game: AiTown,
-    now: number,
-    { playerId, conversationId, timestamp },
-  ): Promise<null> => {
-    const conversation = game.conversations.lookup(conversationId);
-    if (!conversation) {
-      throw new Error(`Invalid conversation ID: ${conversationId}`);
-    }
-    if (conversation.finished) {
-      throw new Error(`Conversation is finished: ${conversationId}`);
-    }
-    if (conversation.isTyping && conversation.isTyping.playerId === playerId) {
-      delete conversation.isTyping;
-    }
-    conversation.lastMessage = { author: playerId, timestamp };
-    conversation.numMessages++;
-    return null;
-  },
-});
-
 export const inputs = {
   join,
   leave,
   moveTo,
-  startConversation,
-  acceptInvite,
-  rejectInvite,
-  leaveConversation,
-  startTyping,
-  finishSendingMessage,
+  // Inputs for the messaging layer.
+  ...conversationInputs,
+  ...conversationMembersInputs,
   // Inputs for the agent layer.
   ...agentInputs,
 };
